Created AuthorRowMapper Write annotation to AuthorValidator, Service,Dao

+++ server/src/main/java/com/getjavajobs/library/services/AuthorService.java
@@ -4,8 +4,10 @@ import com.getjavajobs.library.dao.AuthorDao;
 import com.getjavajobs.library.exceptions.DAOException;
 import com.getjavajobs.library.exceptions.ServiceException;
 import com.getjavajobs.library.model.Author;
+import com.getjavajobs.library.services.validators.AuthorValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -19,43 +21,51 @@ public class AuthorService {
 
     public AuthorService(){};
 
+    @Autowired
+    private AuthorValidator authorValidator;
+
     public AuthorService(AuthorDao authorDao) {
         this.authorDao = authorDao;
     }
 
+    @Transactional
     public Author add(Author author) throws ServiceException {
-        try {
-            return authorDao.add(author);
-        } catch (DAOException e) {
-            throw new ServiceException(e);
+         if(authorValidator.validate()) {
+             return authorDao.add(author);
+         }else{
+            throw new ServiceException("Fall at authorAdd");
         }
     }
+    @Transactional
     public Author get(int id) throws ServiceException {
-        try {
+        if(authorValidator.validate()) {
             return authorDao.get(id);
-        } catch (DAOException e) {
-            throw new ServiceException(e);
+        }else{
+            throw new ServiceException("Fall at authorGet");
         }
     }
+    @Transactional
     public Author update(Author author) throws ServiceException {
-        try {
+        if(authorValidator.validate()) {
             return authorDao.update(author);
-        } catch (DAOException e) {
-            throw new ServiceException(e);
+        }else{
+            throw new ServiceException("Fall at authorUpdate");
         }
     }
+    @Transactional
     public void delete(int id) throws ServiceException{
-        try {
+        if(authorValidator.validate()) {
             authorDao.delete(id);
-        } catch (DAOException e) {
-            throw new ServiceException(e);
+        }else{
+            throw new ServiceException("Fall at authorDelete");
         }
     }
+    @Transactional
     public List<Author> getAll() throws ServiceException{
-        try {
+        if(authorValidator.validate()) {
             return authorDao.getAll();
-        } catch (DAOException e) {
-            throw new ServiceException(e);
+        }else{
+            throw new ServiceException("Fall at authorGetAll");
         }
     }
